Remove AutoLayout::nav_next
This was implemented inconsistently by macros, and in some
cases the correct impl cannot be generated.
